Feature/hyeonwoo 최종

+++ mydictionary_team/src/main/java/com/saltlux/mydictionary/common/Pagination.java
@@ -142,7 +142,7 @@ public class Pagination {
 
 		// 전체 페이지수
 
-		this.pageCnt = (int) Math.ceil(listCnt / listSize);
+		this.pageCnt = (int) Math.ceil((double)listCnt/listSize);
 
 		// 시작 페이지
